remove link_directory_metadata.py
